+++ magda-web-client/src/Pages/HomePage.js
@@ -13,26 +13,22 @@ import { Small, Medium } from "../UI/Responsive";
 
 import MediaQuery from "react-responsive";
 
-const getBgImg = () => {
+const getBgImg = backgroundImageUrls => {
     let imageMap = {};
-    if (!config.homePageConfig || !config.homePageConfig.backgroundImageUrls)
-        return null;
-    const backgroundImageUrls = config.homePageConfig.backgroundImageUrls;
-    if (!backgroundImageUrls.length) return null;
-    const baseUrl = config.homePageConfig.baseUrl
-        ? config.homePageConfig.baseUrl
-        : "";
 
     backgroundImageUrls.forEach(item => {
         let width;
         try {
-            width = parseInt(item.replace(/[^\d]/g, ""), 10);
+            width = parseInt(
+                item.substr(item.lastIndexOf("/") + 1).replace(/[^\d]/g, ""),
+                10
+            );
             if (isNaN(width)) width = 0;
         } catch (e) {
             width = 0;
         }
 
-        imageMap = Object.assign(imageMap, { [width]: baseUrl + item });
+        imageMap = Object.assign(imageMap, { [width]: item });
     });
 
     const screenSizes = Object.keys(imageMap);
@@ -67,50 +63,106 @@ const getBgImg = () => {
     );
 };
 
-const getTagLine = () => {
-    const homePageConfig = config.homePageConfig;
-    return {
-        desktop:
-            config && homePageConfig.tagLineTextDesktop
-                ? homePageConfig.tagLineTextDesktop
-                : "",
-        mobile:
-            config && homePageConfig.tagLineTextMobile
-                ? homePageConfig.tagLineTextMobile
-                : ""
-    };
-};
-
-const HomePage = withRouter(({ location, isTopBannerShown }) => {
-    return (
-        <div className="homepage-app-container">
-            {getBgImg()}
-            <Header />
-            <Small>
-                <div className="homepage-background" />
-            </Small>
-            <div className="container app-container" id="content">
+class HomePage extends React.Component {
+    render() {
+        return (
+            <div className="homepage-app-container">
+                {getBgImg(this.props.backgroundImageUrls)}
+                <Header />
                 <Small>
-                    <TagLine taglineText={getTagLine().mobile} />
+                    <div className="homepage-background" />
                 </Small>
-                <SearchBoxSwitcher location={location} theme="home" />
-                <Medium>
-                    <TagLine taglineText={getTagLine().desktop} />
-                    <Lozenge />
-                </Medium>
-                <Stories />
+                <div className="container app-container" id="content">
+                    {this.props.mobileTagLine && (
+                        <Small>
+                            <TagLine taglineText={this.props.mobileTagLine} />
+                        </Small>
+                    )}
+                    <SearchBoxSwitcher
+                        location={this.props.location}
+                        theme="home"
+                    />
+                    {this.props.desktopTagLine && (
+                        <Medium>
+                            <TagLine taglineText={this.props.desktopTagLine} />
+                            <Lozenge content={this.props.lozenge} />
+                        </Medium>
+                    )}
+                    {(this.props.stories &&
+                        this.props.stories.length && (
+                            <Stories stories={this.props.stories} />
+                        )) ||
+                        ""}
+                </div>
             </div>
-        </div>
-    );
-});
+        );
+    }
+}
 
 function mapStateToProps(state) {
+    let desktopTagLine = "";
+    let mobileTagLine = "";
+    let highlights = {};
+    let stories = {};
+    if (state.content.isFetched) {
+        for (const item of state.content.content) {
+            if (item.id === "home/tagline/desktop") {
+                desktopTagLine = item.content;
+            } else if (item.id === "home/tagline/mobile") {
+                mobileTagLine = item.content;
+            } else if (item.id.indexOf("home/highlights/") === 0) {
+                const id = item.id.substr("home/highlights/".length);
+                highlights[id] = highlights[id] || {};
+                highlights[id].lozenge = item.content;
+            } else if (item.id.indexOf("home/highlight-images/") === 0) {
+                let id = item.id.substr("home/highlight-images/".length);
+                id = id.substr(0, id.lastIndexOf("/"));
+                highlights[id] = highlights[id] || {};
+                highlights[id].backgroundImageUrls =
+                    highlights[id].backgroundImageUrls || [];
+                highlights[id].backgroundImageUrls.push(
+                    `${config.contentApiURL}/${item.id}.bin`
+                );
+            } else if (item.id.indexOf("home/stories/") === 0) {
+                const id = item.id.substr("home/stories/".length);
+                stories[id] = stories[id] || { id };
+                stories[id].content = item.content;
+            } else if (item.id.indexOf("home/story-images/") === 0) {
+                const id = item.id.substr("home/story-images/".length);
+                stories[id] = stories[id] || { id };
+                stories[id].image = `${config.contentApiURL}/${item.id}.bin`;
+            }
+        }
+    }
+    if (Object.keys(highlights).length === 0) {
+        highlights.default = {
+            backgroundImageUrls: [
+                "/assets/homepage/0w.jpg",
+                "/assets/homepage/720w.jpg",
+                "/assets/homepage/1080w.jpg",
+                "/assets/homepage/1440w.jpg",
+                "/assets/homepage/2160w.jpg"
+            ]
+        };
+    }
+    let highlight = Object.keys(highlights);
+    highlight = highlight[new Date().getDate() % highlight.length];
+
+    stories = Object.values(stories)
+        .filter(story => story.content)
+        .sort((a, b) => a.content.order - b.content.order);
+
     return {
-        isTopBannerShown: state.topBanner.isShown
+        isTopBannerShown: state.topBanner.isShown,
+        desktopTagLine,
+        mobileTagLine,
+        lozenge: highlights[highlight].lozenge,
+        backgroundImageUrls: highlights[highlight].backgroundImageUrls,
+        stories
     };
 }
 
 export default connect(
     mapStateToProps,
     null
-)(HomePage);
+)(withRouter(HomePage));