+++ magda-web-client/src/Pages/HomePageComponents/Stories.js
@@ -4,7 +4,6 @@ import { Medium, Small } from "../../UI/Responsive";
 import "./Stories.css";
 import downArrow from "../../assets/downArrow-homepage-more-stories.svg";
 import { CSSTransition, TransitionGroup } from "react-transition-group";
-import { config } from "../../config";
 
 class Stories extends Component {
     constructor(props) {
@@ -30,13 +29,13 @@ class Stories extends Component {
     }
 
     render() {
-        if (
-            !config ||
-            !config.homePageConfig ||
-            !config.homePageConfig.stories ||
-            !config.homePageConfig.stories.length
-        ) {
-            return <div />;
+        const { stories } = this.props;
+        if (!stories || !stories.length) {
+            return null;
+        }
+        const rows = [];
+        for (let x = 0; x < stories.length; x += 4) {
+            rows.push(stories.slice(x, x + 4));
         }
         return (
             <div className="homepage-stories">
@@ -52,46 +51,31 @@ class Stories extends Component {
                             >
                                 <div className="row" key="stories-boxes">
                                     <div className="col-md-12">
-                                        <StoryBox
-                                            idx={0}
-                                            className="story-box-1"
-                                        />
-                                        <StoryBox
-                                            idx={1}
-                                            className="story-box-2"
-                                        />
-                                        <StoryBox
-                                            idx={2}
-                                            className="story-box-3"
-                                        />
-                                        <StoryBox
-                                            idx={3}
-                                            className="story-box-4"
-                                        />
-                                        <StoryBox
-                                            idx={4}
-                                            className="story-box-5"
-                                        />
-                                        <StoryBox
-                                            idx={5}
-                                            className="story-box-6"
-                                        />
+                                        {stories.map((story, index) => (
+                                            <StoryBox
+                                                idx={index}
+                                                story={story}
+                                                className="story-box-{index}"
+                                            />
+                                        ))}
                                     </div>
                                 </div>
                             </CSSTransition>
                         </TransitionGroup>
                     ) : (
-                        <button
-                            className="homepage-stories-show-stories-button"
-                            onClick={() => {
-                                this.setState({
-                                    shouldShowStories: true
-                                });
-                            }}
-                        >
-                            <div>Latest News</div>
-                            <img src={downArrow} alt="more-stories" />
-                        </button>
+                        <div className="homepage-stories-show-stories-button-container">
+                            <button
+                                className="homepage-stories-show-stories-button"
+                                onClick={() => {
+                                    this.setState({
+                                        shouldShowStories: true
+                                    });
+                                }}
+                            >
+                                <div>Latest News</div>
+                                <img src={downArrow} alt="more-stories" />
+                            </button>
+                        </div>
                     )}
                 </Small>
                 <Medium>
@@ -104,34 +88,77 @@ class Stories extends Component {
                             timeout={{ appear: 500 }}
                         >
                             <div>
-                                <div className="stories-container">
-                                    <div className="col-3">
-                                        <StoryBox idx={0} className="stories" />
-                                    </div>
-                                    <div className="col-3">
-                                        <div className="row-2">
-                                            <StoryBox
-                                                idx={1}
-                                                className="row stories"
-                                            />
-                                            <StoryBox
-                                                idx={2}
-                                                className="row stories"
-                                            />
-                                        </div>
-                                    </div>
-                                    <div className="col-3">
-                                        <StoryBox idx={3} className="stories" />
-                                    </div>
-                                </div>
-                                <div className="stories-container">
-                                    <div className="col-2">
-                                        <StoryBox idx={4} className="stories" />
-                                    </div>
-                                    <div className="col-2">
-                                        <StoryBox idx={5} className="stories" />
-                                    </div>
-                                </div>
+                                {rows.map((row, r) => {
+                                    switch (row.length) {
+                                        case 1:
+                                            return (
+                                                <div
+                                                    className="row"
+                                                    key="stories-boxes"
+                                                >
+                                                    <div className="col-md-4">
+                                                        <StoryBox
+                                                            idx={r * 4}
+                                                            story={row[0]}
+                                                            className="stories"
+                                                        />
+                                                    </div>
+                                                </div>
+                                            );
+                                        case 2:
+                                        case 3:
+                                            return (
+                                                <div className="stories-container">
+                                                    {row.map((story, i) => (
+                                                        <div
+                                                            className={`col-${
+                                                                row.length
+                                                            }`}
+                                                        >
+                                                            <StoryBox
+                                                                idx={r * 4 + i}
+                                                                story={story}
+                                                                className="stories"
+                                                            />
+                                                        </div>
+                                                    ))}
+                                                </div>
+                                            );
+                                        default:
+                                            return (
+                                                <div className="stories-container">
+                                                    <div className="col-3">
+                                                        <StoryBox
+                                                            idx={r * 4 + 0}
+                                                            story={row[0]}
+                                                            className="stories"
+                                                        />
+                                                    </div>
+                                                    <div className="col-3">
+                                                        <div className="row-2">
+                                                            <StoryBox
+                                                                idx={r * 4 + 1}
+                                                                story={row[1]}
+                                                                className="row stories"
+                                                            />
+                                                            <StoryBox
+                                                                idx={r * 4 + 2}
+                                                                story={row[2]}
+                                                                className="row stories"
+                                                            />
+                                                        </div>
+                                                    </div>
+                                                    <div className="col-3">
+                                                        <StoryBox
+                                                            idx={r * 4 + 3}
+                                                            story={row[3]}
+                                                            className="stories"
+                                                        />
+                                                    </div>
+                                                </div>
+                                            );
+                                    }
+                                })}
                             </div>
                         </CSSTransition>
                     </TransitionGroup>